+++ Makefile
@@ -7,7 +7,7 @@ help:
 	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' Makefile | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-10s\033[0m %s\n", $$1, $$2}'
 
 bin: ## build the binary
-	$(GO) build -o bin/ocpsel *.go
+	$(GO) build -o bin/ocpsel cmd/ocpsel/*.go
 
 test: ## test
 	$(GO) test -coverprofile c.out ./...
