+++ cmd/ocpsel/main.go
@@ -5,8 +5,9 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"strconv"
 	"strings"
+
+	"github.com/moxspec/moxspec-ocpsel/ocpsel"
 )
 
 func main() {
@@ -23,9 +24,9 @@ func main() {
 			flag.Usage()
 			os.Exit(1)
 		}
-		res, err := decode(sens, gens, eds)
+		res, err := ocpsel.Decode(sens, gens, eds)
 		if err == nil {
-			fmt.Println(res.print())
+			fmt.Println(res.Summary())
 			os.Exit(0)
 		}
 
@@ -40,8 +41,8 @@ func main() {
 		os.Exit(1)
 	}
 
-	chunkList := parseBody(bod)
-	result := parseChunkList(chunkList)
+	chunkList := ocpsel.ParseText(bod)
+	result := ocpsel.ParseChunkList(chunkList)
 
 	if outFile != "" {
 		err := ioutil.WriteFile(outFile, []byte(strings.Join(result, "\n")), 0644)
@@ -53,46 +54,4 @@ func main() {
 	}
 
 	fmt.Println(strings.Join(result, "\n"))
-}
-
-func splitEventData(ed uint64) (byte, byte, byte) {
-	ed1 := byte((ed >> 16) & 0xFF)
-	ed2 := byte((ed >> 8) & 0xFF)
-	ed3 := byte(ed & 0xFF)
-	return ed1, ed2, ed3
-}
-
-func decode(sens, gens, eds string) (record, error) {
-	sen, err := strconv.ParseUint(sens, 16, 32)
-	if err != nil {
-		return record{}, err
-	}
-
-	gen, err := strconv.ParseUint(gens, 16, 32)
-	if err != nil {
-		return record{}, err
-	}
-
-	ed, err := strconv.ParseUint(eds, 16, 32)
-	if err != nil {
-		return record{}, err
-	}
-
-	title, d, err := getDecoder(sen, gen)
-	if err != nil {
-		return record{}, err
-	}
-
-	ed1, ed2, ed3 := splitEventData(ed)
-	r1, r2, r3 := d(ed1, ed2, ed3)
-
-	return record{
-		title: title,
-		ed1:   ed1,
-		r1:    r1,
-		ed2:   ed2,
-		r2:    r2,
-		ed3:   ed3,
-		r3:    r3,
-	}, nil
 }
